% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_document.R
\name{ari_document}
\alias{ari_document}
\title{Ari Document Class for Rendering}
\usage{
ari_document(...)
}
\arguments{
\item{...}{arguments to pass to \link[ari:ari_spin]{ari::ari_spin}}
}
\description{
Ari Document Class for Rendering
}
