% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ari.R
\name{gs_to_ari}
\alias{gs_to_ari}
\alias{pptx_to_ari}
\alias{pdf_to_ari}
\title{Convert Google Slides and notes to video with ari}
\usage{
gs_to_ari(path, ...)

pptx_to_ari(path, script = NULL, ...)

pdf_to_ari(path, dpi = 300, ...)
}
\arguments{
\item{path}{Identifier of google slides presentation, or PPTX filename}

\item{...}{Arguments passed to \link{make_ari_document}}

\item{script}{passed to \code{\link[=make_ari_document]{make_ari_document()}}}

\item{dpi}{resolution (dots per inch) to render images}
}
\value{
The output from \link{make_ari_document}
}
\description{
Convert Google Slides and notes to video with ari
}
