% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_notes.R
\name{pptx_notes}
\alias{pptx_notes}
\title{Get Notes from a PowerPoint (usually from Google Slides)}
\usage{
pptx_notes(file)
}
\arguments{
\item{file}{Character. Path for \code{PPTX} file}
}
\value{
Either a character vector or \code{NULL}
}
\description{
Get Notes from a PowerPoint (usually from Google Slides)
}
\examples{
file = ex_file = system.file("extdata", "example.pptx",
package = "ariExtra")
pptx_notes(ex_file)
}
