% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_notes.R
\name{xml_notes}
\alias{xml_notes}
\title{Get Notes from XML}
\usage{
xml_notes(file, collapse_text = TRUE)
}
\arguments{
\item{file}{XML file from a PPTX}

\item{collapse_text}{should text be collapsed by spaces?}
}
\value{
A character vector
}
\description{
Get Notes from XML
}
