% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_ari.R
\name{rmd_to_ari}
\alias{rmd_to_ari}
\title{Convert a Rmd to an Ari Document}
\usage{
rmd_to_ari(path, script = NULL, capture_method = c("vectorized",
  "iterative"), webshot_args = list(), ..., verbose = TRUE)
}
\arguments{
\item{path}{path to Rmd file}

\item{script}{optional spoken script, otherwise taken from the
HTML comments}

\item{capture_method}{Either \code{"vectorized"} or \code{"iterative"}. The
vectorized mode is faster though it can cause screens to repeat.
If making a video from an \link[rmarkdown:ioslides_presentation]{rmarkdown::ioslides_presentation}
you should use \code{"iterative"}.}

\item{webshot_args}{a list of arguments to pass to \link[webshot:webshot]{webshot::webshot}}

\item{...}{additional arguments to pass to \link{make_ari_document}}

\item{verbose}{print diagnostic messages}
}
\value{
The output of \link{make_ari_document}
}
\description{
Convert a Rmd to an Ari Document
}
\examples{
\dontrun{
path = system.file("extdata", "example.Rmd", package = "ariExtra")
res = rmd_to_ari(path)
}
}
